+++ Source/SwiftLintFramework/Models/PrimaryRuleList.swift
@@ -173,6 +173,7 @@ public let primaryRuleList = RuleList(rules: [
     StaticOperatorRule.self,
     StrictFilePrivateRule.self,
     StrongIBOutletRule.self,
+    StructuredFunctionDocRule.self,
     SuperfluousDisableCommandRule.self,
     SwitchCaseAlignmentRule.self,
     SwitchCaseOnNewlineRule.self,