+++ CHANGELOG.md
@@ -137,6 +137,12 @@
   [Marcelo Fabri](https://github.com/marcelofabri)
   [#3950](https://github.com/realm/SwiftLint/issues/3950)
 
+* Add opt-in `structured_function_doc` rule to enforce markup
+  documentation for functions with a short summary and parameters
+  section matching function declaration.
+  [Maxim Grabarnik](https://github.com/arangato)
+  [#3801](https://github.com/realm/SwiftLint/issues/3801)
+
 #### Bug Fixes
 
 * Fix false positives in `unused_closure_parameter` when using parameters with